requirements.txt
# YieldSwarm AI - Requirements
# ASI Alliance Hackathon Project

# Core ASI Alliance Technologies
uagents>=0.18.0
uagents-core>=0.2.0
hyperon>=0.2.8

# Web3 & Blockchain
web3>=6.0.0
eth-account>=0.10.0
solana>=0.30.0
solders>=0.18.0

# DeFi Protocol Integration
aiohttp>=3.9.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data & Analytics
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
pydantic>=2.0.0
python-dateutil>=2.8.0
